feat(manager): moved api calls to parent component and refactored analytics component
