+++ packages/manager/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@spaship/manager",
   "description": "The SPAship UI.",
-  "version": "2.1.8",
+  "version": "2.1.9",
   "private": true,
   "author": {
     "name": "Akhil Mohan",
