+++ packages/manager/src/services/analytics/index.tsx
@@ -1,7 +1,7 @@
 export {
   useGetDeploymentCounts,
   useGetWebPropActivityStream,
-  useGetMonthlyDeploymentChart,
+  useGetMonthlyDeploymentChartWithEphemeral,
   useGetTotalDeploymentsForApps,
   useGetTotalDeployments,
   useGetDeploymentsTime
