--- packages/manager/src/views/DashboardPage/components/Analytics/components/LineChart/index.tsx
@@ -1,1 +1,0 @@
-export { LineChart } from './LineChart';
