--- packages/manager/src/views/DashboardPage/components/Analytics/components/StatsGrid/index.tsx
@@ -1,1 +1,0 @@
-export { StatsGrid } from './StatsGrid';
